# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "一只轻量级无头CMS",
        "title": "AzuLLia",
        "contact": {},
        "version": "1.0.1206"
    },
    "host": "127.0.0.1:8080",
    "basePath": "/",
    "paths": {
        "/about": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "系统"
                ],
                "summary": "获取框架版本信息",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/service.GeneralJSONHeader"
                        }
                    }
                }
            }
        },
        "/data/article/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文章"
                ],
                "summary": "获取文章内容",
                "parameters": [
                    {
                        "type": "string",
                        "description": "文章ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/service.GeneralJSONHeader"
                        }
                    }
                }
            }
        },
        "/data/category/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "分类"
                ],
                "summary": "获取分类",
                "parameters": [
                    {
                        "type": "string",
                        "description": "分类ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/service.GeneralJSONHeader"
                        }
                    }
                }
            }
        },
        "/data/tag/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "标签"
                ],
                "summary": "获取标签",
                "parameters": [
                    {
                        "type": "string",
                        "description": "标签ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/service.GeneralJSONHeader"
                        }
                    }
                }
            }
        },
        "/list/article": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文章"
                ],
                "summary": "获取列表",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/service.GeneralJSONHeader"
                        }
                    }
                }
            }
        },
        "/list/category": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "分类"
                ],
                "summary": "获取分类列表",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/service.GeneralJSONHeader"
                        }
                    }
                }
            }
        },
        "/list/tag": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "标签"
                ],
                "summary": "获取标签列表",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/service.GeneralJSONHeader"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "service.GeneralJSONHeader": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "object",
                    "additionalProperties": true
                },
                "msg": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                }
            }
        }
    }
}